# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountHealth": {
        "description": "账号健康状态。",
        "enum": [
          "UNKNOWN",
          "HEALTHY",
          "DEGRADED",
          "QUARANTINED"
        ],
        "title": "AccountHealth",
        "type": "string"
      },
...
      "AccountHealthVerification": {
        "additionalProperties": false,
        "description": "A health check together with its final safe account projection.",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/TestAccount"
          },
          "check": {
            "$ref": "#/components/schemas/AccountHealthCheck"
          }
        },
        "required": [
          "check",
          "account"
        ],
        "title": "AccountHealthVerification",
        "type": "object"
      },
...
      "AccountStateReason": {
        "additionalProperties": false,
        "description": "隔离和恢复必须保留结构化安全原因。",
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "AccountStateReason",
        "type": "object"
      },
      "AccountStateTransition": {
        "additionalProperties": false,
        "description": "Immutable before-and-after snapshot of orthogonal account state.",
        "properties": {
          "accountId": {
            "format": "uuid",
            "title": "Accountid",
            "type": "string"
          },
          "actorId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actorid"
          },
          "environmentId": {
            "format": "uuid",
            "title": "Environmentid",
            "type": "string"
          },
          "fromCooldownUntil": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fromcooldownuntil"
          },
          "fromHealthStatus": {
            "$ref": "#/components/schemas/AccountHealth"
          },
          "fromLifecycleStatus": {
            "$ref": "#/components/schemas/AccountLifecycle"
          },
          "fromOperationalStatus": {
            "$ref": "#/components/schemas/AccountOperationalStatus"
          },
          "fromSyncStatus": {
            "$ref": "#/components/schemas/AccountSyncStatus"
          },
          "healthCheckId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Healthcheckid"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "projectId": {
            "format": "uuid",
            "title": "Projectid",
            "type": "string"
          },
          "reason": {
            "$ref": "#/components/schemas/AccountStateTransitionReason"
          },
          "requestId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Requestid",
            "type": "string"
          },
          "safeSummary": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Safesummary",
            "type": "string"
          },
          "tenantId": {
            "format": "uuid",
            "title": "Tenantid",
            "type": "string"
          },
          "toCooldownUntil": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tocooldownuntil"
          },
          "toHealthStatus": {
            "$ref": "#/components/schemas/AccountHealth"
          },
          "toLifecycleStatus": {
            "$ref": "#/components/schemas/AccountLifecycle"
          },
          "toOperationalStatus": {
            "$ref": "#/components/schemas/AccountOperationalStatus"
          },
          "toSyncStatus": {
            "$ref": "#/components/schemas/AccountSyncStatus"
          }
        },
        "required": [
          "id",
          "tenantId",
          "projectId",
          "environmentId",
          "accountId",
          "healthCheckId",
          "reason",
          "fromLifecycleStatus",
          "toLifecycleStatus",
          "fromHealthStatus",
          "toHealthStatus",
          "fromOperationalStatus",
          "toOperationalStatus",
          "fromSyncStatus",
          "toSyncStatus",
          "fromCooldownUntil",
          "toCooldownUntil",
          "safeSummary",
          "actorId",
          "requestId",
          "occurredAt"
        ],
        "title": "AccountStateTransition",
        "type": "object"
      },
      "AccountStateTransitionPage": {
        "additionalProperties": false,
        "description": "Cursor page of AccountStateTransition facts.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AccountStateTransition"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "AccountStateTransitionPage",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Atlas 扩展的 RFC Problem Details。",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errorCode": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            },
            "title": "Violations",
            "type": "array"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "errorCode",
          "requestId"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "TestAccount": {
        "additionalProperties": false,
        "description": "不包含任何秘密的被测系统账号投影。",
        "properties": {
          "accountKey": {
            "title": "Accountkey",
            "type": "string"
          },
          "authMethods": {
            "items": {
              "$ref": "#/components/schemas/CredentialAuthMethod"
            },
            "title": "Authmethods",
            "type": "array"
          },
          "availabilityReason": {
            "$ref": "#/components/schemas/AccountAvailabilityReason"
          },
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "connectorInstallationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connectorinstallationid"
          },
          "consecutiveHealthFailures": {
            "minimum": 0.0,
            "title": "Consecutivehealthfailures",
            "type": "integer"
          },
          "cooldownUntil": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldownuntil"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "credentialValid": {
            "title": "Credentialvalid",
            "type": "boolean"
          },
          "environmentId": {
            "format": "uuid",
            "title": "Environmentid",
            "type": "string"
          },
          "externalSubjectId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Externalsubjectid"
          },
          "healthStatus": {
            "$ref": "#/components/schemas/AccountHealth"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "labels": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Labels",
            "type": "object"
          },
          "lastHealthCheckedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lasthealthcheckedat"
          },
          "lastHealthSucceededAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lasthealthsucceededat"
          },
          "lastLeasedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastleasedat"
          },
          "leaseEpoch": {
            "minimum": 0.0,
            "title": "Leaseepoch",
            "type": "integer"
          },
          "lifecycleStatus": {
            "$ref": "#/components/schemas/AccountLifecycle"
          },
          "loginHintMasked": {
            "title": "Loginhintmasked",
            "type": "string"
          },
          "operationalStatus": {
            "$ref": "#/components/schemas/AccountOperationalStatus"
          },
          "poolId": {
            "format": "uuid",
            "title": "Poolid",
            "type": "string"
          },
          "projectId": {
            "format": "uuid",
            "title": "Projectid",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/AccountSource"
          },
          "syncStatus": {
            "$ref": "#/components/schemas/AccountSyncStatus"
          },
          "tenantId": {
            "format": "uuid",
            "title": "Tenantid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "projectId",
          "environmentId",
          "poolId",
          "connectorInstallationId",
          "accountKey",
          "source",
          "externalSubjectId",
          "loginHintMasked",
          "lifecycleStatus",
          "healthStatus",
          "operationalStatus",
          "syncStatus",
          "cooldownUntil",
          "consecutiveHealthFailures",
          "lastHealthCheckedAt",
          "lastHealthSucceededAt",
          "leaseEpoch",
          "labels",
          "lastLeasedAt",
          "authMethods",
          "credentialValid",
          "available",
          "availabilityReason",
          "revision",
          "createdAt",
          "updatedAt"
        ],
        "title": "TestAccount",
        "type": "object"
      },
...
      "VerifyTestAccount": {
        "additionalProperties": false,
        "description": "Request a login and role health check on one exact Origin.",
        "properties": {
          "origin": {
            "maxLength": 2048,
            "minLength": 8,
            "title": "Origin",
            "type": "string"
          }
        },
        "required": [
          "origin"
        ],
        "title": "VerifyTestAccount",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Atlas TestOps Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/test-accounts/{accountId}/state-transitions": {
      "get": {
        "description": "Return immutable before-and-after snapshots of orthogonal account state.",
        "operationId": "list_test_account_state_transitions_v1_test_accounts__accountId__state_transitions_get",
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Accountid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountStateTransitionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求语义或 Revision 无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "权限、Production 或 Origin 策略拒绝"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "TestAccount 不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "账号状态、Lease 或并发检查冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "TestAccount Revision 已变化"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Secret Provider、Adapter 或能力不可用"
          }
        },
        "summary": "列出 TestAccount 状态迁移",
        "tags": [
          "account-health"
        ]
      }
    },
    "/v1/test-accounts/{accountId}:quarantine": {
      "post": {
        "operationId": "quarantine_test_account_v1_test_accounts__accountId__quarantine_post",
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Accountid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "maxLength": 64,
              "title": "If-Match",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountStateReason"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestAccount"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求语义或 Revision 无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "当前 PlatformRole 无权执行"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "唯一键、幂等或状态转换冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Revision 前置条件失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          }
        },
        "summary": "隔离 TestAccount",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/test-accounts/{accountId}:restore": {
      "post": {
        "operationId": "restore_test_account_v1_test_accounts__accountId__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Accountid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "maxLength": 64,
              "title": "If-Match",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountStateReason"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestAccount"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求语义或 Revision 无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "当前 PlatformRole 无权执行"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "唯一键、幂等或状态转换冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Revision 前置条件失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          }
        },
        "summary": "恢复 TestAccount",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/test-accounts/{accountId}:verify": {
      "post": {
        "description": "Run an out-of-transaction Provider probe and apply it with revision CAS.",
        "operationId": "verify_test_account_v1_test_accounts__accountId__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Accountid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "maxLength": 64,
              "title": "If-Match",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTestAccount"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountHealthVerification"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求语义或 Revision 无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "权限、Production 或 Origin 策略拒绝"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "TestAccount 不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "账号状态、Lease 或并发检查冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "TestAccount Revision 已变化"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Secret Provider、Adapter 或能力不可用"
          }
        },
        "summary": "验证 TestAccount 登录身份与角色健康",
        "tags": [
          "account-health"
        ]
      }
    },
...
  }
}